+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Install rbenv
 - Automatically add default SSH key, generating if required
 - Stop it repeatedly updating Homebrew
+### Changed
+- Handle casks AKA Mac Applications better
 ### Fixed
 - Issue with `gtc` command when running from a repository's subdirectory
 ### Removed